+++ tests/utils/test_mimebundle.py
@@ -14,6 +14,11 @@ try:
 except ImportError:
     vlc = None
 
+try:
+    import vegafusion as vf  # type: ignore
+except ImportError:
+    vf = None
+
 
 @pytest.fixture
 def vegalite_spec():
@@ -242,6 +247,7 @@ def check_pre_transformed_vega_spec(vega_spec):
     assert len(data_0.get("transform", [])) == 0
 
 
+@pytest.mark.skipif(vf is None, reason="vegafusion is not installed")
 def test_vegafusion_spec_to_vega_mime_bundle(vegalite_spec):
     with alt.data_transformers.enable("vegafusion"):
         bundle = spec_to_mimebundle(
@@ -254,6 +260,7 @@ def test_vegafusion_spec_to_vega_mime_bundle(vegalite_spec):
         check_pre_transformed_vega_spec(vega_spec)
 
 
+@pytest.mark.skipif(vf is None, reason="vegafusion is not installed")
 def test_vegafusion_chart_to_vega_mime_bundle(vegalite_spec):
     chart = alt.Chart.from_dict(vegalite_spec)
     with alt.data_transformers.enable("vegafusion"), alt.renderers.enable("json"):